refactor expression analyzer
