+++ server/src/main/java/org/apache/iotdb/db/mpp/plan/expression/visitor/cartesian/ConcatExpressionWithSuffixPathsVisitor.java
@@ -17,16 +17,17 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.mpp.plan.expression.visitor;
+package org.apache.iotdb.db.mpp.plan.expression.visitor.cartesian;
 
-import org.apache.iotdb.commons.path.MeasurementPath;
 import org.apache.iotdb.commons.path.PartialPath;
-import org.apache.iotdb.db.mpp.common.schematree.ISchemaTree;
+import org.apache.iotdb.commons.path.PathPatternTree;
+import org.apache.iotdb.db.constant.SqlConstant;
 import org.apache.iotdb.db.mpp.plan.expression.Expression;
 import org.apache.iotdb.db.mpp.plan.expression.leaf.ConstantOperand;
 import org.apache.iotdb.db.mpp.plan.expression.leaf.TimeSeriesOperand;
 import org.apache.iotdb.db.mpp.plan.expression.leaf.TimestampOperand;
 import org.apache.iotdb.db.mpp.plan.expression.multi.FunctionExpression;
+import org.apache.iotdb.tsfile.common.constant.TsFileConstant;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -35,65 +36,77 @@ import java.util.List;
 import static org.apache.iotdb.db.mpp.plan.analyze.ExpressionUtils.cartesianProduct;
 import static org.apache.iotdb.db.mpp.plan.analyze.ExpressionUtils.reconstructFunctionExpressions;
 import static org.apache.iotdb.db.mpp.plan.analyze.ExpressionUtils.reconstructTimeSeriesOperands;
-import static org.apache.iotdb.db.utils.TypeInferenceUtils.bindTypeForAggregationNonSeriesInputExpressions;
 
-public class RemoveWildcardInExpressionVisitor extends CartesianProductVisitor<ISchemaTree> {
+public class ConcatExpressionWithSuffixPathsVisitor
+    extends CartesianProductVisitor<ConcatExpressionWithSuffixPathsVisitor.Context> {
   @Override
   public List<Expression> visitFunctionExpression(
-      FunctionExpression functionExpression, ISchemaTree schemaTree) {
-    // One by one, remove the wildcards from the input expressions. In most cases, an expression
-    // will produce multiple expressions after removing the wildcards. We use extendedExpressions
-    // to collect the produced expressions.
+      FunctionExpression functionExpression, Context context) {
     List<List<Expression>> extendedExpressions = new ArrayList<>();
-    for (Expression originExpression : functionExpression.getExpressions()) {
-      List<Expression> actualExpressions = process(originExpression, schemaTree);
-      if (actualExpressions.isEmpty()) {
-        // Let's ignore the eval of the function which has at least one non-existence series as
-        // input. See IOTDB-1212: https://github.com/apache/iotdb/pull/3101
-        return Collections.emptyList();
-      }
-      extendedExpressions.add(actualExpressions);
+    for (Expression suffixExpression : functionExpression.getExpressions()) {
+      extendedExpressions.add(process(suffixExpression, context));
 
       // We just process first input Expression of AggregationFunction,
-      // keep other input Expressions as origin and bind Type
+      // keep other input Expressions as origin
       // If AggregationFunction need more than one input series,
       // we need to reconsider the process of it
       if (functionExpression.isBuiltInAggregationFunctionExpression()) {
         List<Expression> children = functionExpression.getExpressions();
-        bindTypeForAggregationNonSeriesInputExpressions(
-            ((FunctionExpression) functionExpression).getFunctionName(),
-            children,
-            extendedExpressions);
+        for (int i = 1; i < children.size(); i++) {
+          extendedExpressions.add(Collections.singletonList(children.get(i)));
+        }
         break;
       }
     }
-
-    // Calculate the Cartesian product of extendedExpressions to get the actual expressions after
-    // removing all wildcards. We use actualExpressions to collect them.
     List<List<Expression>> childExpressionsList = new ArrayList<>();
     cartesianProduct(extendedExpressions, childExpressionsList, 0, new ArrayList<>());
-
-    return reconstructFunctionExpressions(
-        (FunctionExpression) functionExpression, childExpressionsList);
+    return reconstructFunctionExpressions(functionExpression, childExpressionsList);
   }
 
   @Override
   public List<Expression> visitTimeSeriesOperand(
-      TimeSeriesOperand timeSeriesOperand, ISchemaTree schemaTree) {
-    PartialPath path = timeSeriesOperand.getPath();
-    List<MeasurementPath> actualPaths = schemaTree.searchMeasurementPaths(path).left;
+      TimeSeriesOperand timeSeriesOperand, Context context) {
+    PartialPath rawPath = timeSeriesOperand.getPath();
+    List<PartialPath> actualPaths = new ArrayList<>();
+    if (rawPath.getFullPath().startsWith(SqlConstant.ROOT + TsFileConstant.PATH_SEPARATOR)) {
+      actualPaths.add(rawPath);
+      context.getPatternTree().appendPathPattern(rawPath);
+    } else {
+      for (PartialPath prefixPath : context.getPrefixPaths()) {
+        PartialPath concatPath = prefixPath.concatPath(rawPath);
+        context.getPatternTree().appendPathPattern(concatPath);
+        actualPaths.add(concatPath);
+      }
+    }
     return reconstructTimeSeriesOperands(actualPaths);
   }
 
   @Override
   public List<Expression> visitTimeStampOperand(
-      TimestampOperand timestampOperand, ISchemaTree context) {
+      TimestampOperand timestampOperand, Context context) {
     return Collections.singletonList(timestampOperand);
   }
 
   @Override
-  public List<Expression> visitConstantOperand(
-      ConstantOperand constantOperand, ISchemaTree context) {
+  public List<Expression> visitConstantOperand(ConstantOperand constantOperand, Context context) {
     return Collections.singletonList(constantOperand);
+  }
+
+  public static class Context {
+    private final List<PartialPath> prefixPaths;
+    private final PathPatternTree patternTree;
+
+    public Context(List<PartialPath> prefixPaths, PathPatternTree patternTree) {
+      this.prefixPaths = prefixPaths;
+      this.patternTree = patternTree;
+    }
+
+    public List<PartialPath> getPrefixPaths() {
+      return prefixPaths;
+    }
+
+    public PathPatternTree getPatternTree() {
+      return patternTree;
+    }
   }
 }
